refactor(projects): move project hooks to a new directory and update imports
