+++ hooks/project/project-keys.ts
@@ -1,0 +1,8 @@
+export const projectKeys = {
+  all: ['projects'] as const,
+  lists: () => [...projectKeys.all, 'list'] as const,
+  list: (filters?: Record<string, unknown>) =>
+    [...projectKeys.lists(), filters] as const,
+  details: () => [...projectKeys.all, 'detail'] as const,
+  detail: (id: number) => [...projectKeys.details(), id] as const,
+};
